Make the build behavior consistent with the docs.
The docs say we should be able to build an AOSP tree by pointing
LLVM_PREBUILTS_BASE at our stage 2 install directory. Testing this has
proven that to be a lie.

Don't install the final prebuilts to a temporary directory, but to a
well known path of $OUT_DIR/install.

Also, change the default build name away from the date because the
user will also need to override LLVM_PREBUILTS_VERSION, and that would
just be annoying to type. Unless otherwise specified, the final
package name will be "clang-dev".

Change-Id: Ia9ba84ec793e9e2364ab148683ec48aacb19ec4a