Pass SOONG_ALLOW_MISSING_DEPENDENCIES
Soong has stricter dependency checks for platform branches than for
stripped down branches. Switch off the strict checks so that dependency
checking will happen when a module is attempted to be built instead of
when we're generating the build rules.

USE_SOONG no longer does anything, and we've switched to using Kati
prebuilts that should work on 10.8, so remove it here.

Bug: 25715107
Change-Id: Id371660d42694f7628fae4e7c35a4a3cc99b1312