Skip building libomp on Darwin
Bug: http://b/27617302

Darwin fails to build with errors "only weak aliases are supported on
darwin" in file kmp_ftn_entry.h.  It is possible to fix these since
upstream has support to build on Darwin (by running cmake configuration
separately on Darwin).

This is not a high priority (or even a necessity).  So, skip building on
Darwin.

Change-Id: I23556d22a70a9dbad2d69c0f42c40b0c70a10f91