Don't use the compiler wrappers on Windows.
File extensions matter on Windows. We can't invoke clang.exe if
clang.exe is a Python file, and all of the build systems invoke
clang.exe directly, so installing it as clang.py won't do much good
either. Just skip the wrappers for now since we're not using them in
the NDK yet anyway.

Test: ./build.py
Bug: None
Change-Id: Ia2b272bd334e94c5bfc77ac9bb20f8e02e054c0c