Only build 64-bit host for stage1
stage1 only needs the 64-bit host compiler to be used for building
stage2.  Disable building the 32-bit host compiler completely, as it
is never needed.  Depend on a target that only builds for the host,
preventing building the 32-bit and 64-bit windows clang, and prevent
installing windows clang in stage1.

Reduces build output by ~30GB.

Bug: 33464365
Test: external/clang/build.py
Change-Id: I17ee608362254db3c4fbbaed9ea4ce13b4092c4f