Add support for prebuilt libFuzzer.a.
Bug: http://b/35432050

We need to rename libLLVMFuzzer.a to libFuzzer.a to be more consistent
with external usage of libFuzzer (and avoid duplicate definitions for
the libLLVMFuzzer component). We also include .h/.def files from this
directory, as users will automatically want to pull these in to make the
most out of libFuzzer.

Test: external/clang/build.py + manual verification of files/headers.
Change-Id: I55834cf51cf3595ba513b54abbdb8e8851f1f2ba